requirements.txt
pymongo==4.7.3
neo4j==5.20.0
pandas==2.2.2
matplotlib==3.9.0
tabulate==0.9.0
python-dotenv==1.0.1
